+++ R/n_factors.R
@@ -1,0 +1,177 @@
+#' Find Optimal Factor Number.
+#'
+#' Find optimal factor number using various solutions.
+#'
+#' @param df The dataframe
+#' @param rotate What rotation to use c("none", "varimax", "oblimin","promax")
+#' @param fm Factoring method – fm="pa" Principal Axis Factor Analysis,
+#' fm = "minres" minimum residual (OLS) factoring fm="mle" Maximum Likelihood FA,
+#' fm="pc" Principal Components"
+#' @param n_max How many factors to test.
+#'
+#' @return output
+#'
+#' @examples
+#' df <- dplyr::select_if(attitude, is.numeric)
+#' results <- psycho::n_factors(df)
+#'
+#' summary(results)
+#' plot(results)
+#'
+#' # See details on methods
+#' psycho::values(results)$methods
+#'
+#' @author \href{https://dominiquemakowski.github.io/}{Dominique Makowski}
+#'
+#' @importFrom qgraph cor_auto
+#' @importFrom psych VSS
+#' @importFrom MASS mvrnorm
+#' @importFrom MASS ginv
+#' @importFrom nFactors moreStats
+#' @importFrom nFactors nScree
+#' @importFrom stats cov
+#' @importFrom stats dnorm
+#' @importFrom stats qnorm
+#' @export
+n_factors <- function(df, rotate="varimax", fm="minres", n_max=8){
+
+  # Copy the parallel function from nFactors to correct the use of mvrnorm
+  parallel <- function (subject = 100, var = 10, rep = 100, cent = 0.05, quantile = cent,
+            model = "components", sd = diag(1, var), ...)
+  {
+    r <- subject
+    c <- var
+    y <- matrix(c(1:r * c), nrow = r, ncol = c)
+    ycor <- matrix(c(1:c * c), nrow = c, ncol = c)
+    evpea <- NULL
+    leg.txt <- "Pearson"
+    for (k in c(1:rep)) {
+      y <- MASS::mvrnorm(n = r, mu = rep(0, var), Sigma = sd, empirical = FALSE)
+      corY <- cov(y, ...)
+      if (model == "components")
+        diag(corY) <- diag(sd)
+      if (model == "factors")
+        corY <- corY - MASS::ginv(diag(diag(MASS::ginv(corY))))
+      evpea <- rbind(evpea, eigen(corY)[[1]])
+    }
+    SEcentile <- function(sd, n = 100, p = 0.95) {
+      return(sd/sqrt(n) * sqrt(p * (1 - p))/dnorm(qnorm(p)))
+    }
+    sprob <- c(cent)
+    mevpea <- sapply(as.data.frame(evpea), mean)
+    sevpea <- sapply(as.data.frame(evpea), sd)
+    qevpea <- nFactors::moreStats(evpea, quantile = quantile)[3, ]
+    sqevpea <- sevpea
+    sqevpea <- sapply(as.data.frame(sqevpea), SEcentile, n = rep,
+                      p = cent)
+    result <- list(eigen = data.frame(mevpea, sevpea, qevpea,
+                                      sqevpea), subject = r, variables = c, centile = cent)
+    class(result) <- "parallel"
+    return(result)
+  }
+
+
+  cor <- qgraph::cor_auto(df)
+
+  ap <- parallel(subject=nrow(df), var=ncol(df))
+  nS <- nFactors::nScree(x=eigen(cor)$values, aparallel=ap$eigen$qevpea)
+
+  # Eigeinvalues data
+  eigenvalues <- nS$Analysis %>%
+    dplyr::select_("Eigenvalues",
+                  "Exp.Variance"="Prop",
+                  "Cum.Variance"="Cumu") %>%
[…]
