+++ vignettes/overview.Rmd
@@ -8,7 +8,7 @@ author:
 date: "`r Sys.Date()`"
 tags: [r, psychology, neuroscience]
 abstract: |
-  Psycho is an R package that aims at providing tools for psychologists, neuropsychologists and neuroscientists, to transform statistical outputs into something readable that can be, almost directly, copied and pasted into a report. It also implements various functions useful in psychological science, such as correlation matrices, assessment plot creation or normalization. The package revolves around the psychobject. Main functions from the package return this type, and the analyze() function transforms other R objects (for now, only stan_lmer type) into psychobjects. Four functions can then be applied on a psychobject: summary(), print(), plot() and values(). Contrary to many other packages which goal is to produce statistical analyzes, `psycho`'s goal is to fill the gap between statistical R output and statistical report writing, with a focus on APA formatting guidelines. Complex outputs, such as those of Bayesian linear models, are automatically transformed into readable text, important values are extracted and plots are drawn to illustrate the effects. Thus, the results can easily be incorporated into shareable reports and publications, saving time and preventing errors for better, reproducible, science.
+  Psycho is an R package that aims at providing tools for psychologists, neuropsychologists and neuroscientists, to transform statistical outputs into something readable that can be, almost directly, copied and pasted into a report. It also implements various functions useful in psychological science, such as correlation matrices, assessment plot creation or normalization. The package revolves around the psychobject. Main functions from the package return this type, and the `analyze()` function transforms other R objects into psychobjects. Four functions can then be applied on a psychobject: `summary()`, `print()`, `plot()` and `values()`. Contrary to many other packages which goal is to produce statistical analyzes, `psycho` aims at filling the gap between statistical R outputs and statistical report writing, with a focus on APA formatting guidelines, to enhance the standardization of results reporting. Complex outputs, such as those of Bayesian and frequentist mixed models, are automatically transformed into readable text, tables, and plots that illustrate the effects. Thus, the results can easily be incorporated into shareable reports and publications, promoting data exploration, saving time and preventing errors for better, reproducible, science. 
 vignette: >
   %\VignetteIndexEntry{Overview}
   %\VignetteEngine{knitr::rmarkdown}
@@ -149,7 +149,43 @@ print(results)
 plot(results)
 ```
 
+------
 
+## How many factors/components to retain?
+
+The `n_factors()` function is useful in before running principal component (PCA) or factor (FA) analysis. As many statistical methods exists to that purpose, this function gathers them together and gives an overview on the most frequent result. It also draw a nice plot with the eigenvalues and the proportion of explained  variance.
+
+
+```{r echo=FALSE, message=FALSE, warning=FALSE, results='hide'}
+results <- attitude %>%
+  select_if(is.numeric) %>% 
+  psycho::n_factors()
+
+# Get a summary
+summary(results)
+```
+
+```{r echo=FALSE, message=FALSE, warning=FALSE}
+kable(summary(results))
+```
+
+We can also extract the final result (the optimal number of factors) for each method:
+
+```{r echo=FALSE, message=FALSE, warning=FALSE, results='hide'}
+psycho::values(results)$methods
+```
+
+```{r echo=FALSE, message=FALSE, warning=FALSE}
+kable(psycho::values(results)$methods)
+```
+
+And, of course, plot it :)
+
+```{r, fig.width=7, fig.height=4.5, eval = TRUE, results='markup', fig.align='center'}
+plot(results)
+```
+     
+     
 ------
 
 ## Analyze the Mixed-Modelling Framework
