### 2017-11-11 e5366c3
make lintr a bit happier
