### 2017-11-11 1cb4481
n_factors fix
for fm="pc"
