+++ R/correlation.R
@@ -51,11 +51,13 @@ correlation <- function(df,
 
   # Remove non numeric
   df <- df[ , sapply(df, is.numeric)]
-
+  if (is.null(df2) == FALSE) {
+    df2 <- df2[ , sapply(df2, is.numeric)]
+  }
   # Compute r coefficients
   if (type == "full") {
     corr <- psych::corr.test(df, y = df2, use = "pairwise", method = method, adjust="none")
-    r <- corr$p
+    r <- corr$r
     p <- corr$p
     t <- corr$t
     ci <- corr$ci
@@ -72,7 +74,7 @@ correlation <- function(df,
     if (type == "semi") {corr <- ppcor::spcor(df, method = method)}
     else if (type == "partial") {corr <- ppcor::pcor(df, method = method)}
     else {
-      warning("psycho correlation(): type must be 'full', 'semi' or 'partial'")
+      warning("type parameter must be 'full', 'semi' or 'partial'")
       return()
     }
     r <- corr$estimate
@@ -86,8 +88,12 @@ correlation <- function(df,
   # Adjust P values
   n <- nrow(df)
   if (adjust != "none"){
-    p[lower.tri(p)] <- p.adjust(p[lower.tri(p)], method=adjust, n=choose(nrow(p), 2))
-    p[upper.tri(p)] <- p.adjust(p[upper.tri(p)], method=adjust, n=choose(nrow(p), 2))
+    if ((type == "full" & is.null(df2)==F)|(type == "semi")) {
+      p <- p.adjust(p, method=adjust)
+    } else{
+      p[lower.tri(p)] <- p.adjust(p[lower.tri(p)], method=adjust, n=choose(nrow(p), 2))
+      p[upper.tri(p)] <- p.adjust(p[upper.tri(p)], method=adjust, n=choose(nrow(p), 2))
+    }
   }
 
 
@@ -99,17 +105,29 @@ correlation <- function(df,
 
   # trunctuate the matrix that holds the correlations to two decimal
   r_format <- format(round(cbind(rep(-1.11, ncol(df)), r), 2))[,-1]
-  ## build a new correlation matrix with significance stars
+  # build a new correlation matrix with significance stars
   table <- matrix(paste(r_format, mystars, sep = ""), ncol = ncol(df))
-  diag(table) <- paste(diag(r_format), " ", sep = "")
-  rownames(table) <- colnames(df)
-  colnames(table) <- paste(colnames(df), "", sep = "")
-  # remove upper triangle
-  table <- as.matrix(table)
-  table[upper.tri(table, diag = TRUE)] <- NA
-  table <- as.data.frame(table)
-  # remove last column and return the matrix (which is now a data frame)
-  table <- cbind(table[1:length(table) - 1])
+
+  # Format
+  if ((type == "full" & is.null(df2)==F)|(type == "semi")) {
+    if (type == "semi"){
+      rownames(table) <- colnames(df)
+    } else{
+      rownames(table) <- colnames(df2)
+    }
+    colnames(table) <- paste(colnames(df), "", sep = "")
+    table <- as.matrix(table)
+    table <- as.data.frame(table)
+  } else {
+    diag(table) <- paste(diag(r_format), " ", sep = "")
+    rownames(table) <- colnames(df)
+    colnames(table) <- paste(colnames(df), "", sep = "")
+    table <- as.matrix(table)
+    table[upper.tri(table, diag = TRUE)] <- NA  # remove upper triangle
+    table <- as.data.frame(table)
+    table <- cbind(table[1:length(table) - 1])  # remove last column and return the matrix (which is now a data frame)
+  }
+
 
 
   # Values
@@ -127,17 +145,42 @@ correlation <- function(df,
 
   # Plot
   # -------------
[…]
