### 2017-11-16 89a8391
added R2 for stan_lm
