### 2017-11-25 #470 @whedon
[REVIEW]: The psycho Package: an Efficient and Publishing-Oriented Workflow for Psychological Science
**Submitting author:** @DominiqueMakowski (<a href="http://orcid.org/0000-0001-5375-9967">Dominique Makowski</a>)
**Repository:** <a href="https://github.com/neuropsychology/psycho.R" target ="_blank">https://github.com/neuropsychology/psycho.R</a>
**Version:** 0.0.2
