### 2017-11-30 fbf3498
hotfix
display median, MAD, CI and MPE in analyze.stanreg
