+++ R/analyze.stanreg.R
@@ -109,9 +109,9 @@ analyze.stanreg <- function(x, CI=95, effsize=FALSE, verbose=T, ...){
                   format_digit(MPE_values[1]), " and ",
                   format_digit(MPE_values[2]),
                   " (Median = ",format_digit(median),
-                  # ", MAD = ", format_digit(mad),
+                  ", MAD = ", format_digit(mad),
                   # ", Mean = ", format_digit(mean),
-                  ", SD = ", format_digit(sd),
+                  # ", SD = ", format_digit(sd),
                   ", ", CI, "% CI [",
                   format_digit(CI_values[1]), ", ",
                   format_digit(CI_values[2]), "], ",
