+++ NAMESPACE
@@ -12,6 +12,8 @@ export(correlation)
 export(find_season)
 export(format_digit)
 export(format_p)
+export(format_string)
+export(hdi)
 export(interpret_d)
 export(n_factors)
 export(normalize)
