+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: psycho
 Type: Package
 Title: Efficient and Publishing-Oriented Workflow for Psychological Science
-Version: 0.0.5
+Version: 0.0.6
 Authors@R: c(
     person("Dominique", 
         "Makowski", 
