+++ R/format_p.R
@@ -10,5 +10,9 @@ format_p <- function(pvalues) {
     ifelse(pvalues < 0.01, "< .01**",
       ifelse(pvalues < 0.05, "< .05*",
         ifelse(pvalues < 0.1, paste0("= ", round(pvalues, 2), "\xB0"),
-          "> .1"))))
+          "> .1"
+        )
+      )
+    )
+  )
 }
