+++ R/formatting.R
@@ -8,14 +8,33 @@
 #'
 #' @export
 format_digit <- function(x, digits = 2, null_treshold = 0.001) {
-  rounded <- if (abs(x) < 1) {
-    signif(x, digits)
-  } else {
-    round(x, digits)
-  }
 
-  if (abs(rounded) < null_treshold) "0" else rounded
+  # If x is an Integer
+  if (x %% 1 == 0) {
+    formatted <- as.character(x)
+  } else {
+    # If x is close to zero
+    if (abs(x) < null_treshold) {
+      formatted <- "0"
+    } else {
+      # If x is close to trailing zeros
+      if (abs(x) < 1) {
+        formatted <- as.character(signif(x, digits))
+        # If signif cut off trailing zero, add it
+        # TODO: that line of code is ugly
+        if (nchar(gsub("0|-|\\.", "", formatted)) < digits) {
+          formatted <- paste0(formatted, strrep("0", digits - 1))
+        }
+      } else {
+        formatted <- format_string(round(x, digits), paste0("%.", digits, "f"))
+      }
+    }
+  }
+  return(formatted)
 }
+
+
+
 
 
 
@@ -24,8 +43,7 @@ format_digit <- function(x, digits = 2, null_treshold = 0.001) {
 #'
 #' @param x Values.
 #' @param fmt A character vector of format strings, each of up to 8192 bytes.
-#' @param ... values to be passed into fmt. Only logical, integer, real and character vectors are supported, but some coercion will be done: see the ‘Details’ section. Up to 100.
-#'
+
 #'
 #' @export
 format_string <- function(x, fmt, ...) {
