+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: psycho
 Type: Package
 Title: Efficient and Publishing-Oriented Workflow for Psychological Science
-Version: 0.0.6
+Version: 0.0.7
 Authors@R: c(
     person("Dominique", 
         "Makowski", 
