### 2018-01-04 3cf818d
0.0.7b
Added except to normalize
