+++ man/normalize.Rd
@@ -4,10 +4,12 @@
 \alias{normalize}
 \title{Normalize (scale and reduce) numeric variables.}
 \usage{
-normalize(df)
+normalize(df, except = NULL)
 }
 \arguments{
 \item{df}{Dataframe.}
+
+\item{except}{Character or list of characters of column names to be excluded from normalization.}
 }
 \value{
 Dataframe.
@@ -16,11 +18,17 @@ Dataframe.
 Select numeric variables and normalize (Z-score) them.
 }
 \examples{
-df <- data.frame(Participant = as.factor(rep(1:50,each=2)),
-  Condition = base::rep_len(c("A", "B"), 100), V1 = rnorm(100, 30, .2),
-  V2 = runif(100, 3, 5))
+df <- data.frame(
+  Participant = as.factor(rep(1:50,each=2)),
+  Condition = base::rep_len(c("A", "B"), 100),
+  V1 = rnorm(100, 30, .2),
+  V2 = runif(100, 3, 5),
+  V3 = rnorm(100, 100, 10)
+  )
 
 dfZ <- normalize(df)
+dfZ <- normalize(df, except="V3")
+dfZ <- normalize(df, except=c("V1", "V2"))
 
 }
 \author{
