### 2018-01-10 #470 @SachaEpskamp
Hi, sorry for the delay! This looks like a very nice small package that could be useful in several ways. The installation and examples all perform as expected. Below are some observations I had in no particular order.

Best, Sacha

1. Just entering an object in the console does not print it:
data("bfi")
cres <- correlation(bfi[,1:25])
cres # does nothing
print(cres) # prints results

2. It is not clear where I can find and/or if there are arguments to the print method to, say, the output of correlation( )

3. I don't really like the NA's printed in the correlation output, would be nicer if that was an empty string by default?

4. It would be nice if polychoric and polyserial correlations could be included. The qgraph function cor_auto detects values and uses the lavaan function lavCor to obtain polychoric/polyserial correlations. Perhaps something like this could be included in correlation( ) as well?

5. I like the inclusion of partial correlations. Perhaps an option could be to estimate these using the glasso package rather than significance testing? the EBICglasso function from qgraph, which wraps around glasso from the glasso package, should easily be useable to this end in the backend.

6. It's not clear from the documentation what, if any, arguments I can use in the plot method. For example, can I change the colors of the assess plot (which is really nice btw).

7. The documentation should make clearer what parts are based on other packages. For example, n_factors uses qgraph's cor_auto in the background, which uses lavaan in the background, but both packages are not referenced. I think the psycho package also heavily relies on the psych package, but it is not referenced in the documentation at all.

8. I'd suggest to set the argument forcePD = FALSE in cor_auto. It was a bad design choice of me to make that default to TRUE, and has lead to a lot of problems.

9. Perhaps EGA (DOI:10.1371/journal.pone.0174035) can be included in n_factors?

10. I like the normalize() function, but don't really like the name, as it implies it will make data normally distributed (as the huge.npn() function in huge for example). The function standardizes, not normalizes. Perhaps it could be renamed to standardize?

10. The analyze() function is nice, but I don't really like the string labeling of effect sizes as "Very small", as I am always in favor of numeric effect size measures rather than  labels based on arbitrary thresholds. Can't the numeric effect size be printed instead?

11. When printing the output of analyze( ), I would make use of the cat( ) function to present the output nicely instead of the print function. That is, instead of:

> print(results)
[1] "The overall model predicting ... successfully converged and explained 10.57% of the variance of the endogen (the conditional R2). The variance explained by the fixed effects was of 0.56% (the marginal R2) and the one explained by the random effects of 10.01%."
[2] "The effect of (Intercept) was [NOT] significant (beta = 0.092, SE = 0.22, t(5.81) = 0.42, p > .1) and can be considered as very small (std. beta = 0, std. SE = 0)."
[3] "The effect of ConditionB was [NOT] significant (beta = -0.15, SE = 0.20, t(95.00) = -0.78, p > .1) and can be considered as very small (std. beta = -0.076, std. SE = 0.097)."

I think the following would be nicer:

> print(results)
The overall model predicting ... successfully converged and explained 10.57% of the variance of the endogen (the conditional R2). The variance explained by the fixed effects was of 0.56% (the marginal R2) and the one explained by the random effects of 10.01%.

	- The effect of (Intercept) was [NOT] significant (beta = 0.092, SE = 0.22, t(5.81) = 0.42, p > .1) and can be considered as very small (std. beta = 0, std. SE = 0).
	- The effect of ConditionB was [NOT] significant (beta = -0.15, SE = 0.20, t(95.00) = -0.78, p > .1) and can be considered as very small (std. beta = -0.076, std. SE = 0.097).

Perhaps the invisible() function can be used to still return the character vector with these texts as well so that the function can be used in a .Rnw or .Rmd file to generate a report though.
