### 2018-01-10 #470 @DominiqueMakowski
Wow, thanks @SachaEpskamp for these very useful comments! (altough some of which will require from me some more learning on programming and statistics 😄).

I'll open an issue for each one of them to track progress and try to work out as many as I can.

@leeper, I'll come back here once I've given my best to let you know of the advances.

Thanks again ☺️
