### 2018-01-10 neuropsychology/psycho.R#21 @DominiqueMakowski
JOSS Review - 1 - Just entering an object in the console does not print it
Point 1 from [JOSS Review](https://github.com/openjournals/joss-reviews/issues/470#issuecomment-356727181):

--------------

Just entering an object in the console does not print it:
```r
data("bfi")
cres <- correlation(bfi[,1:25])
cres # does nothing
print(cres) # prints results
```
