### 2018-01-10 neuropsychology/psycho.R#22 @DominiqueMakowski
JOSS Review - 2 - arguments to the print method
Point 2 from [JOSS Review](https://github.com/openjournals/joss-reviews/issues/470#issuecomment-356727181):

--------------

It is not clear where I can find and/or if there are arguments to the print method to, say, the output of correlation( )
