### 2018-01-10 neuropsychology/psycho.R#23 @DominiqueMakowski
JOSS Review - 3 - NA's printed in the correlation output
Point 3 from [JOSS Review](https://github.com/openjournals/joss-reviews/issues/470#issuecomment-356727181):

--------------

I don't really like the NA's printed in the correlation output, would be nicer if that was an empty string by default?
