### 2018-01-10 neuropsychology/psycho.R#24 @DominiqueMakowski
JOSS Review - 4 - polychoric and polyserial correlations
Point 4 from [JOSS Review](https://github.com/openjournals/joss-reviews/issues/470#issuecomment-356727181):

--------------

It would be nice if polychoric and polyserial correlations could be included. The qgraph function cor_auto detects values and uses the lavaan function lavCor to obtain polychoric/polyserial correlations. Perhaps something like this could be included in correlation( ) as well?
