### 2018-01-10 neuropsychology/psycho.R#25 @DominiqueMakowski
JOSS Review - 5 - partial correlations with glasso
Point 5 from [JOSS Review](https://github.com/openjournals/joss-reviews/issues/470#issuecomment-356727181):

--------------

I like the inclusion of partial correlations. Perhaps an option could be to estimate these using the glasso package rather than significance testing? the EBICglasso function from qgraph, which wraps around glasso from the glasso package, should easily be useable to this end in the backend.
