### 2018-01-10 neuropsychology/psycho.R#26 @DominiqueMakowski
JOSS Review - 6 - plot method: arguments and personalization
Point 6 from [JOSS Review](https://github.com/openjournals/joss-reviews/issues/470#issuecomment-356727181):

--------------

It's not clear from the documentation what, if any, arguments I can use in the plot method. For example, can I change the colors of the assess plot (which is really nice btw).
