### 2018-01-10 neuropsychology/psycho.R#27 @DominiqueMakowski
JOSS Review - 7 - Dependencies
Point 7 from [JOSS Review](https://github.com/openjournals/joss-reviews/issues/470#issuecomment-356727181):

--------------

The documentation should make clearer what parts are based on other packages. For example, n_factors uses qgraph's cor_auto in the background, which uses lavaan in the background, but both packages are not referenced. I think the psycho package also heavily relies on the psych package, but it is not referenced in the documentation at all.
