### 2018-01-10 neuropsychology/psycho.R#28 @DominiqueMakowski
JOSS Review - 8 - cor_auto parameter
Point 8 from [JOSS Review](https://github.com/openjournals/joss-reviews/issues/470#issuecomment-356727181):

--------------

I'd suggest to set the argument forcePD = FALSE in cor_auto. It was a bad design choice of me to make that default to TRUE, and has lead to a lot of problems.
