### 2018-01-10 neuropsychology/psycho.R#29 @DominiqueMakowski
JOSS Review - 9 - include EGA in n_factors
Point 9 from [JOSS Review](https://github.com/openjournals/joss-reviews/issues/470#issuecomment-356727181):

--------------

Perhaps [EGA](http://journals.plos.org/plosone/article?id=10.1371/journal.pone.0174035) can be included in n_factors?
