### 2018-01-10 neuropsychology/psycho.R#30 @DominiqueMakowski
JOSS Review - 10 - normalize to standardize
Point 10 from [JOSS Review](https://github.com/openjournals/joss-reviews/issues/470#issuecomment-356727181):

--------------

I like the normalize() function, but don't really like the name, as it implies it will make data normally distributed (as the huge.npn() function in huge for example). The function standardizes, not normalizes. Perhaps it could be renamed to standardize?
