### 2018-01-10 neuropsychology/psycho.R#31 @DominiqueMakowski
JOSS Review - 11 -  numeric effect size measures rather than labels based on arbitrary thresholds
Point 11 from [JOSS Review](https://github.com/openjournals/joss-reviews/issues/470#issuecomment-356727181):

--------------

The analyze() function is nice, but I don't really like the string labeling of effect sizes as "Very small", as I am always in favor of numeric effect size measures rather than labels based on arbitrary thresholds. Can't the numeric effect size be printed instead?
