### 2018-01-10 neuropsychology/psycho.R#32 @DominiqueMakowski
[This](http://adv-r.had.co.nz/S3.html) might be related to that issue:

"If you’re implementing more complicated print() methods, it’s a better idea to implement format() methods that return a string, and then implement print.class <- function(x, ...) cat(format(x, ...), "\n". This makes for methods that are much easier to compose, because the side-effects are isolated to a single place."
