### 2018-01-11 neuropsychology/psycho.R#21 @DominiqueMakowski
I've [asked](https://stackoverflow.com/questions/48197304/how-to-add-a-default-behaviour-to-an-s3-class?noredirect=1#comment83378269_48197304) how to fix that but was unsuccessful (apparently this issue is not consistent across users?) 😞
