### 2018-01-11 neuropsychology/psycho.R#25 @DominiqueMakowski
This has been added in upcomming release.
