### 2018-01-11 neuropsychology/psycho.R#30 @DominiqueMakowski
I agree, this has been fixed in upcomming version.
