### 2018-01-11 neuropsychology/psycho.R#24 @DominiqueMakowski
I've added the possibility to use `cor_auto` instead of `psych::corr.test`. Unfortunately, cor_auto doesn't return p values (altough I know it's not a bad thing **per se**), so it cannot fully replace the default method for now.
