+++ R/correlation.R
@@ -5,8 +5,11 @@
 #' @param df The dataframe
 #' @param df2 Optional dataframe to correlate with the first one.
 #' @param type A character string indicating which correlation type is to be
-#'   computed. One of "full" (default), "partial" or "semi" for semi-partial
-#'   correlations.
+#'   computed. One of "full" (default), "partial" (partial correlations),
+#'   "semi" (semi-partial correlations), "glasso"
+#'   (Graphical lasso- estimation of Gaussian graphical models) or "cor_auto"
+#'   (will use the qgraph::cor_auto function to return pychoric or polyserial
+#'   correlations if needed).
 #' @param method A character string indicating which correlation coefficient is
 #'   to be computed. One of "pearson" (default), "kendall", or "spearman" can be
 #'   abbreviated.
@@ -34,9 +37,10 @@
 #'
 #' @author \href{https://dominiquemakowski.github.io/}{Dominique Makowski}
 #'
-#' @importFrom stats na.omit p.adjust
+#' @importFrom stats na.omit p.adjust cor
 #' @importFrom psych corr.test
 #' @importFrom ggplot2 theme element_text
+#' @importFrom stringr str_to_title
 #' @import ggcorrplot
 #' @import ppcor
 #' @export
@@ -48,6 +52,8 @@ correlation <- function(df,
 
   # Processing
   # -------------------
+  # N samples
+  n <- nrow(df)
 
   # Remove non numeric
   df <- df[, sapply(df, is.numeric)]
@@ -72,80 +78,203 @@ correlation <- function(df,
     df <- stats::na.omit(df) # enable imputation
     if (type == "semi") {
       corr <- ppcor::spcor(df, method = method)
+      r <- corr$estimate
+      p <- corr$p.value
+      t <- corr$statistic
+      ci <- "Not available for partial and semipartial correlations."
+      ci.adj <- "Not available for partial and semipartial correlations."
     }
     else if (type == "partial") {
       corr <- ppcor::pcor(df, method = method)
+      r <- corr$estimate
+      p <- corr$p.value
+      t <- corr$statistic
+      ci <- "Not available for partial and semipartial correlations."
+      ci.adj <- "Not available for partial and semipartial correlations."
+    }
+    else if (type == "glasso") {
+      corr <- qgraph::EBICglasso(cor(df), n, gamma = 0.5)
+      r <- corr
+      p <-  NULL
+      t <- NULL
+      ci <- "Not available for glasso estimation."
+      ci.adj <- "Not available for glasso estimation."
+    }
+    else if (type == "cor_auto") {
+      corr <- qgraph::cor_auto(df, forcePD = F)
+      r <- corr
+      p <-  NULL
+      t <- NULL
+      ci <- "Not available for cor_auto estimation."
+      ci.adj <- "Not available for cor_auto estimation."
     }
     else {
-      warning("type parameter must be 'full', 'semi' or 'partial'")
+      warning("type parameter must be 'full', 'semi', 'partial', 'glasso' or 'cor_auto'")
       return()
     }
-    r <- corr$estimate
-    p <- corr$p.value
-    t <- corr$statistic
-    ci <- "Not available for partial and semipartial correlations."
-    ci.adj <- "Not available for partial and semipartial correlations."
   }
+
 
 
   # Adjust P values
-  n <- nrow(df)
-  if (adjust != "none") {
-    if ((type == "full" & is.null(df2) == F) | (type == "semi")) {
-      p <- p.adjust(p, method = adjust)
-    } else {
[…]
