+++ R/standardize.R
@@ -1,6 +1,6 @@
-#' Normalize (scale and reduce) numeric variables.
+#' Standardize (scale and reduce) numeric variables.
 #'
-#' Select numeric variables and normalize (Z-score) them.
+#' Select numeric variables and standardize (Z-score, "normalize") them.
 #'
 #' @param df Dataframe.
 #' @param except Character or list of characters of column names to be excluded from normalization.
@@ -16,9 +16,9 @@
 #'   V3 = rnorm(100, 100, 10)
 #'   )
 #'
-#' dfZ <- normalize(df)
-#' dfZ <- normalize(df, except="V3")
-#' dfZ <- normalize(df, except=c("V1", "V2"))
+#' dfZ <- standardize(df)
+#' dfZ <- standardize(df, except="V3")
+#' dfZ <- standardize(df, except=c("V1", "V2"))
 #'
 #' @author \href{https://dominiquemakowski.github.io/}{Dominique Makowski}
 #'
@@ -26,7 +26,7 @@
 #' @import purrr
 #' @import dplyr
 #' @export
-normalize <- function(df, except=NULL) {
+standardize <- function(df, except=NULL) {
 
   # Remove exceptions
   if(!is.null(except) && except %in% names(df)){
