+++ tests/testthat/test-standardize.R
@@ -1,4 +1,4 @@
-context("normalize")
+context("standardize")
 
 test_that("Correct Value", {
   df <- data.frame(
@@ -6,7 +6,7 @@ test_that("Correct Value", {
     V2 = runif(100, 3, 5),
     V3 = rnorm(100, 100, 10)
   )
-  dfZ <- normalize(df)
+  dfZ <- standardize(df)
   testthat::expect_equal(mean(dfZ$V1), 0)
 
   df <- data.frame(
@@ -16,9 +16,9 @@ test_that("Correct Value", {
     V2 = runif(100, 3, 5),
     V3 = rnorm(100, 100, 10)
     )
-  dfZ <- normalize(df, except="V3")
+  dfZ <- standardize(df, except="V3")
   testthat::expect_equal(mean(dfZ$V2), 0)
-  dfZ <- normalize(df, except=c("V1", "V2"))
+  dfZ <- standardize(df, except=c("V1", "V2"))
   testthat::expect_equal(mean(dfZ$V3), 0)
 
 })
