+++ tests/testthat/test-is.psychobject.R
@@ -1,0 +1,8 @@
+context("is.psychobject")
+
+test_that("is.psychobject", {
+  df <- attitude
+  results <- psycho::correlation(df)
+  testthat::expect_true(psycho::is.psychobject(results))
+
+})
