### 2018-01-11 5e2b6ff
Adding EGA to n_factors
