### 2018-01-11 neuropsychology/psycho.R#29 @DominiqueMakowski
That would be a really nice addition.

However, I've [tried] to add it, but 1) the [EGA](https://github.com/hfgolino/EGA) package doens't seem available yet on CRAN, and I couldn't successfuly run the functions with the github version.

I'm commenting the EGA-related lines in n_factors for now, until a miracle happens.
