### 2018-01-11 neuropsychology/psycho.R#32 @DominiqueMakowski
Great idea!
This has been fixed in upcomming release.
