### 2018-01-11 3b95f82
nicer analyze output
