+++ R/analyze.merModLmerTest.R
@@ -113,7 +113,7 @@ analyze.merModLmerTest <- function(x, ...) {
   ))
 
   for (varname in varnames) {
-    text <- c(text, values[[varname]]$Text)
+    text <- c(text, paste("   -", values[[varname]]$Text))
   }
 
 
