### 2018-01-11 69876b8
Added credits section to readme
