### 2018-01-11 neuropsychology/psycho.R#27 @DominiqueMakowski
I added a [credit section](https://github.com/neuropsychology/psycho.R/commit/69876b863a341e9098ae0f8267033786f5c3ac82) to readme. Is that sufficient?
