### 2018-01-11 neuropsychology/psycho.R#21 @DominiqueMakowski
Seems to be fixed in new version.
