### 2018-01-11 neuropsychology/psycho.R#22 @DominiqueMakowski
There are no additional argument for the `print.psychobject` method (the one used by `correlation()`, as it returns  a "psychobject").
Indeed, that method only prints the text contained in the psychobject, which itself is computed at the psychobject-creation level (through functions such as `correlation()`, `analyze()` or `assess()`).

However, in general, arguments can be found in the [documentation](https://www.rdocumentation.org/packages/psycho/versions/0.0.8/topics/print.psychobject) ☺️ .
