### 2018-01-11 neuropsychology/psycho.R#26 @DominiqueMakowski
I've added some information about custom plots in the vignette.
