+++ vignettes/overview.Rmd
@@ -75,11 +75,11 @@ cor <- psycho::correlation(df,
                            method = "pearson",
                            adjust = "none")
 
-print(cor)
+summary(cor)
 ```
 
 ```{r echo=FALSE, message=FALSE, warning=FALSE}
-kable(print(cor))
+kable(summary(cor))
 ```
 
 
@@ -95,9 +95,9 @@ plot(cor)
 ### Partial, Corrected, Correlations
 
 `correlation()` offers the possibility to run partial or semi-partial 
-correleations.
+correleations, as well as printing them pairwise.
 
-```{r, fig.width=8, eval = TRUE, fig.align='center', results='hide'}
+```{r, fig.width=8, eval = TRUE, fig.align='center', results='markup', comment=NA}
 library(psycho)
 
 df <- iris
@@ -110,9 +110,7 @@ pcor <- psycho::correlation(df,
 print(pcor)
 ```
 
-```{r echo=FALSE, message=FALSE, warning=FALSE}
-kable(print(pcor))
-```
+
 
 
 ------
@@ -121,7 +119,7 @@ kable(print(pcor))
 
 The `standardize()` function allows you to easily scale and center all numeric variables of a dataframe. It is similar to the base function `scale()`, but presents some advantages: it is tidyverse-friendly, data-type friendly (*i.e.*, does not transform it into a matrix) and can handle dataframes with categorical data.
 
-```{r, out.width=8, eval = TRUE, fig.align='center', results='markup'}
+```{r, out.width=8, eval = TRUE, fig.align='center', results='markup', comment=NA}
 library(psycho)
 library(tidyverse)
 
@@ -131,6 +129,7 @@ iris %>%
   summary()
 ```
 
+
 ------
 
 ## Assess
@@ -138,7 +137,7 @@ iris %>%
 This function is useful in clinical activity. It is sometimes necessary to show to the patient, his family or other members of staff, a visual representation of his score. The `assess()` function also computes the percentile and the Z-score, often needed for neuropsychological reports.
 
 
-```{r, fig.width=7, fig.height=4.5, eval = TRUE, results='markup', fig.align='center'}
+```{r, fig.width=7, fig.height=4.5, eval = TRUE, results='markup', fig.align='center', comment=NA}
 library(psycho)
 
 results <- psycho::assess(124, mean=100, sd=15)
@@ -149,6 +148,49 @@ print(results)
 # Plot it
 plot(results)
 ```
+
+It is also possible to custom the plot a bit (see the [documentation](https://www.rdocumentation.org/packages/psycho/versions/0.0.8/topics/assess) for available parameters).
+
+```{r, fig.width=7, fig.height=4.5, eval = TRUE, results='markup', fig.align='center'}
+library(psycho)
+
+results <- psycho::assess(85, mean=100, sd=15, linecolor = "orange", fillcolor = "#4CAF50")
+
+# Plot it
+plot(results)
+```
+
+
+
+------
+
+## Custom Plots
+
+In general, the `plot()` function returns, most of the times, a ggplot object. That means it remains quite flexible. Here's an example.
+
+
+```{r eval=TRUE, fig.align='center', fig.height=4.5, fig.width=9, message=FALSE, warning=FALSE, results='markup'}
+library(psycho)
+
+# Let's create a correlation plot
+p <- plot(psycho::correlation(iris))
+
+# Custom theme and colours
+p <- p + 
+  scale_fill_gradientn(colors = c("#4CAF50", "#FFEB3B", "#FF5722")) +
+  ylab("Variables\n") +
+  labs(fill = "r") +
+  theme(plot.background = element_rect(fill = "#607D8B"),
+        axis.title.y = element_text(size = 20, angle = 90, colour="white"),
[…]
