+++ tests/testthat/test-is.psychobject.R
@@ -4,5 +4,4 @@ test_that("is.psychobject", {
   df <- attitude
   results <- psycho::correlation(df)
   testthat::expect_true(psycho::is.psychobject(results))
-
 })
