+++ R/print.psychobject.R
@@ -8,6 +8,6 @@
 #' @export
 print.psychobject <- function(x, ...) {
   text <- x$text
-  cat(text, sep="\n")
+  cat(text, sep = "\n")
   invisible(text)
 }
