### 2018-01-11 9a877d5
Remove lintr comments
You can drown in these comments.
