+++ .travis.yml
@@ -15,4 +15,4 @@ r_packages:
 
 after_success:
   - Rscript -e 'library(covr); codecov()'
-  - Rscript -e 'devtools::install();lintr::lint_package()'
+#  - Rscript -e 'devtools::install();lintr::lint_package()'
