### 2018-01-11 9eeeb21
Paper improvement
