### 2018-01-11 #470 @DominiqueMakowski
@leeper I've tackled as many points as I could, and left three open (waiting for opinion or help). @SachaEpskamp Again, thanks a lot for your great suggestions 😌
