### 2018-01-12 d41c341
Update paper.md
