### 2018-01-12 615dc9d
Update paper.md
