### 2018-01-14 d65410e
init
Add merMod support
