+++ NAMESPACE
@@ -1,6 +1,7 @@
 # Generated by roxygen2: do not edit by hand
 
 S3method(analyze,glmerMod)
+S3method(analyze,merMod)
 S3method(analyze,merModLmerTest)
 S3method(analyze,stanreg)
 S3method(plot,psychobject)
@@ -33,6 +34,7 @@ importFrom(MuMIn,r.squaredGLMM)
 importFrom(MuMIn,std.coef)
 importFrom(ggplot2,element_text)
 importFrom(ggplot2,theme)
+importFrom(methods,slot)
 importFrom(nFactors,moreStats)
 importFrom(nFactors,nScree)
 importFrom(psych,VSS)
