### 2018-02-04 #470 @leeper
@DominiqueMakowski I'm happy with this as is. Can you please generate a DOI archive for the software in its current form (e.g., using Zenodo) and let me know the DOI?
