+++ vignettes/overview.R
@@ -26,15 +26,15 @@ cor <- psycho::correlation(df,
                            method = "pearson",
                            adjust = "none")
 
-print(cor)
+summary(cor)
 
 ## ----echo=FALSE, message=FALSE, warning=FALSE----------------------------
-kable(print(cor))
+kable(summary(cor))
 
 ## ---- fig.width=8, eval = TRUE, fig.align='center'-----------------------
 plot(cor)
 
-## ---- fig.width=8, eval = TRUE, fig.align='center', results='hide'-------
+## ---- fig.width=8, eval = TRUE, fig.align='center', results='markup', comment=NA----
 library(psycho)
 
 df <- iris
@@ -44,21 +44,21 @@ pcor <- psycho::correlation(df,
                            method = "pearson",
                            adjust = "bonferroni")
 
+summary(pcor)
+
+## ---- results='markup', comment=NA---------------------------------------
 print(pcor)
 
-## ----echo=FALSE, message=FALSE, warning=FALSE----------------------------
-kable(print(pcor))
-
-## ---- out.width=8, eval = TRUE, fig.align='center', results='markup'-----
+## ---- out.width=8, eval = TRUE, fig.align='center', results='markup', comment=NA----
 library(psycho)
 library(tidyverse)
 
 iris %>% 
   select(Species, Sepal.Length, Petal.Length) %>% 
-  psycho::normalize() %>% 
+  psycho::standardize() %>% 
   summary()
 
-## ---- fig.width=7, fig.height=4.5, eval = TRUE, results='markup', fig.align='center'----
+## ---- fig.width=7, fig.height=4.5, eval = TRUE, results='markup', fig.align='center', comment=NA----
 library(psycho)
 
 results <- psycho::assess(124, mean=100, sd=15)
@@ -69,7 +69,35 @@ print(results)
 # Plot it
 plot(results)
 
-## ----echo=FALSE, message=FALSE, warning=FALSE, results='hide'------------
+## ---- fig.width=7, fig.height=4.5, eval = TRUE, results='markup', fig.align='center'----
+library(psycho)
+
+results <- psycho::assess(85, mean=100, sd=15, linecolor = "orange", fillcolor = "#4CAF50")
+
+# Plot it
+plot(results)
+
+## ----eval=TRUE, fig.align='center', fig.height=4.5, fig.width=9, message=FALSE, warning=FALSE, results='markup'----
+library(psycho)
+
+# Let's create a correlation plot
+p <- plot(psycho::correlation(iris))
+
+# Custom theme and colours
+p <- p + 
+  scale_fill_gradientn(colors = c("#4CAF50", "#FFEB3B", "#FF5722")) +
+  ylab("Variables\n") +
+  labs(fill = "r") +
+  theme(plot.background = element_rect(fill = "#607D8B"),
+        axis.title.y = element_text(size = 20, angle = 90, colour="white"),
+        axis.text = element_text(size = 15, colour="white"),
+        legend.title = element_text(size = 20, colour="white"),
+        legend.text = element_text(size = 15, colour="white"),
+        title = element_text(size = 16, colour="white"))
+p
+
+
+## ----echo=TRUE, message=FALSE, warning=FALSE, results='hide'-------------
 results <- attitude %>%
   select_if(is.numeric) %>% 
   psycho::n_factors()
@@ -98,8 +126,8 @@ df <- data.frame(Participant = as.factor(rep(1:25, each = 4)),
                  RT = rnorm(100, 30, .2), 
                  Stress = runif(100, 3, 5))
 
-# Normalize the numeric variables.
-df <- psycho::normalize(df)
+# Standardize the numeric variables.
+df <- psycho::standardize(df)
 
 # Take a look  at the first 10 rows
[…]
