+++ NEWS.md
@@ -1,4 +1,4 @@
-# Current Dev 0.0.9
+# Current Dev 0.1.0
 
 ### Breaking changes
 - `normalize` has been renamed to `standardize` #30
@@ -8,6 +8,7 @@
 - Added glasso and cor_auto estimation for `correlation` #25 #24
 - added `is.psychobject` function.
 ### Changes
+- CRAN ready
 - Fixed #28
 - Nicer print.psychobject output #32
 - Replaced NAs by empty strings in `correlation` #23
