### 2018-02-05 eccce8d
Update NEWS.md
