### 2018-02-05 04dca5f
Update NEWS.md
