### 2018-02-05 #470 @DominiqueMakowski
@leeper The DOI is the following: 10.5281/zenodo.1165620

Thanks a lot
