### 2018-02-05 #470 @DominiqueMakowski
I added the paper link in the readme,

Thanks again everyone!! 🎉
